+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.4.1] - 2018-12-07
+
+- 0.4.0 changes (sequence numbers, derived values when `build()` called without any arguments) added to async as well as sync.
+- Changed makeFactory to take a config object instead of just a starting sequence number, so that any future factory-level config is not an API breaking change
+
 ## [0.4.0] - 2018-12-06
 
 ### Changed